src/ui/requirements.txt
chainlit>=2.0.1
python-dotenv>=1.0.1
dapr>=1.14.0
fastapi>=0.115.6
uvicorn>=0.34.0
./libs/vanilla_aiagents-3.0.0-py3-none-any.whl[remote]